[WRDS-18](rename): Custom -> Primitive, RoundButton -> ChipButton
